+++ packages/compat/webpack/src/types/config/tools.ts
@@ -3,10 +3,12 @@ import type {
   ChainedConfig,
   FileFilterUtil,
   SharedToolsConfig,
-  ToolsLessConfig,
-  ToolsSassConfig,
   TerserPluginOptions,
 } from '@rsbuild/shared';
+import type {
+  BabelTransformOptions,
+  BabelConfigUtils,
+} from '@rsbuild/plugin-babel';
 import type {
   ModifyWebpackChainUtils,
   ModifyWebpackConfigUtils,
@@ -48,7 +50,19 @@ export type ToolsWebpackChainConfig = ArrayOrNot<
   (chain: WebpackChain, utils: ModifyWebpackChainUtils) => void
 >;
 
+export type ToolsBabelConfig = ChainedConfig<
+  BabelTransformOptions,
+  BabelConfigUtils
+>;
+
 export interface ToolsConfig extends SharedToolsConfig {
+  /**
+   * Modify the options of [babel-loader](https://github.com/babel/babel-loader)
+   * When `tools.babel`'s type is Function，the default babel config will be passed in as the first parameter, the config object can be modified directly, or a value can be returned as the final result.
+   * When `tools.babel`'s type is `Object`, the config will be shallow merged with default config by `Object.assign`.
+   * Note that `Object.assign` is a shallow copy and will completely overwrite the built-in `presets` or `plugins` array, please use it with caution.
+   */
+  babel?: ToolsBabelConfig;
   /**
    * Modify the options of [terser-webpack-plugin](https://github.com/webpack-contrib/terser-webpack-plugin).
    */
