+++ packages/plugin-babel/src/index.ts
@@ -1,15 +1,9 @@
-import { DefaultRsbuildPlugin } from '@rsbuild/shared';
-import { applyRspackBabelConfig } from './rspack';
-import { PluginBabelOptions } from './type';
-
-export const pluginBabel = (
-  options: PluginBabelOptions = {},
-): DefaultRsbuildPlugin => ({
-  name: 'plugin-babel',
-
-  setup(api) {
-    if (api.context.bundlerType === 'rspack') {
-      applyRspackBabelConfig(api, options);
-    }
-  },
-});
+export { pluginBabel } from './plugin';
+export { getBabelUtils } from './helper';
+export type {
+  PresetEnvOptions,
+  PresetEnvTargets,
+  PresetEnvBuiltIns,
+  BabelConfigUtils,
+  BabelTransformOptions,
+} from './types';
