--- packages/plugin-babel/src/type.ts
@@ -1,10 +1,0 @@
-import type {
-  ChainedConfig,
-  BabelTransformOptions,
-  BabelConfigUtils,
-} from '@rsbuild/shared';
-
-export type PluginBabelOptions = ChainedConfig<
-  BabelTransformOptions,
-  BabelConfigUtils
->;
