--- packages/shared/src/applyDefaultRsbuildOptions.ts
@@ -1,11 +1,0 @@
-import { CreateRsbuildOptions } from './types/rsbuild';
-
-export function applyDefaultRsbuildOptions(options?: CreateRsbuildOptions) {
-  return {
-    cwd: process.cwd(),
-    entry: {},
-    target: ['web'],
-    configPath: null,
-    ...options,
-  } as Required<CreateRsbuildOptions>;
-}
