--- packages/shared/src/svgo.ts
@@ -1,17 +1,0 @@
-export function getSvgoDefaultConfig() {
-  return {
-    plugins: [
-      {
-        name: 'preset-default',
-        params: {
-          overrides: {
-            // viewBox is required to resize SVGs with CSS.
-            // @see https://github.com/svg/svgo/issues/1128
-            removeViewBox: false,
-          },
-        },
-      },
-      'prefixIds',
-    ],
-  };
-}
