+++ packages/shared/src/types/config/tools.ts
@@ -1,5 +1,4 @@
 import type { ArrayOrNot, ChainedConfig, FileFilterUtil } from '../utils';
-import { BabelConfigUtils, BabelTransformOptions } from '../../babel';
 import type {
   AutoprefixerOptions,
   SassLoaderOptions,
@@ -77,11 +76,6 @@ export type ToolsBundlerChainConfig = ArrayOrNot<
   (chain: BundlerChain, utils: ModifyBundlerChainUtils) => void
 >;
 
-export type ToolsBabelConfig = ChainedConfig<
-  BabelTransformOptions,
-  BabelConfigUtils
->;
-
 export type ToolsPugConfig = true | ChainedConfig<PugOptions>;
 
 export type ToolsPostCSSLoaderConfig = ChainedConfig<
@@ -116,13 +110,6 @@ export interface SharedToolsConfig {
    * Modify the options of DevServer.
    */
   devServer?: ToolsDevServerConfig;
-  /**
-   * Modify the options of [babel-loader](https://github.com/babel/babel-loader)
-   * When `tools.babel`'s type is Function，the default babel config will be passed in as the first parameter, the config object can be modified directly, or a value can be returned as the final result.
-   * When `tools.babel`'s type is `Object`, the config will be shallow merged with default config by `Object.assign`.
-   * Note that `Object.assign` is a shallow copy and will completely overwrite the built-in `presets` or `plugins` array, please use it with caution.
-   */
-  babel?: ToolsBabelConfig;
   /**
    * Modify the options of [css-loader](https://github.com/webpack-contrib/css-loader).
    */
